Move SortableTable module to separate file
